Function to return either normal, lognormal, gamma, or non-parametric distribution estimate

Function to return either normal, lognormal, gamma, or non-parametric distribution estimate

est_dist_censored(df, left_censored, method = "sf", alpha = 0.05,
  choices = c("norm", "lnorm"), combine_locations = FALSE,
  keep_data_object = FALSE)

Arguments

df
groundwater data frame in tidy format

left_censored
logical vector.

method
default is "sf"

alpha
alpha

choices
vector of distributions to check. Default is c("norm, "lnorm")

combine_locations
TRUE/FALSE to estimate distribution by individual location, or grouped together. Default is FALSE.

keep_data_object
Default is FALSE

Examples

data("gw_data")
wells <- c("MW-1", "MW-2", "MW-3")
params <- c("Fluoride, total",
            "Arsenic, dissolved",
            "Nickel, dissolved")

gw_data <- gw_data %>%
to_censored() %>%
filter(location_id %in% wells, param_name %in% params) %>%
percent_lt() %>%
filter(percent_lt >=15, percent_lt <= 50)

gw_data %>%
group_by(location_id, param_name, default_unit) %>%
est_dist_censored(., alpha = 0.05, keep_data_object = FALSE)

#> Warning: `cols` is now required.
#> Please use `cols = c(distribution)`

#> # A tibble: 6 x 4
#> # Groups:   location_id, default_unit, param_name [6]
#>   location_id default_unit param_name         distribution 
#>   <chr>       <chr>        <chr>              <chr>        
#> 1 MW-2        mg/L         Fluoride, total    Normal       
#> 2 MW-2        ug/L         Arsenic, dissolved Lognormal    
#> 3 MW-1        ug/L         Arsenic, dissolved Nonparametric
#> 4 MW-1        ug/L         Nickel, dissolved  Normal       
#> 5 MW-3        mg/L         Fluoride, total    Nonparametric
#> 6 MW-3        ug/L         Arsenic, dissolved Normal       

gw_data %>%
group_by(param_name, default_unit) %>%
est_dist_censored(., combine_locations = TRUE, keep_data_object = FALSE)

#> Warning: `cols` is now required.
#> Please use `cols = c(distribution)`

#> # A tibble: 3 x 3
#> # Groups:   default_unit, param_name [3]
#>   default_unit param_name         distribution
#>   <chr>        <chr>              <chr>       
#> 1 mg/L         Fluoride, total    Lognormal   
#> 2 ug/L         Arsenic, dissolved Lognormal   
#> 3 ug/L         Nickel, dissolved  Normal