Grubb's test for outliers

Grubb's test for outliers

grubbs_test(df, x = "analysis_result", type = 10, opposite = FALSE,
  two_sided = FALSE, group_by_location = FALSE,
  keep_data_object = FALSE)

Arguments

df
data frame

x
column of analysis results

type
integer value indicating test variant. 10 is a test for one outlier (side is detected automatically and can be reversed by opposite parameter). 11 is a test for two outliers on opposite tails, 20 is test for two outliers in one tail. Default is 10

opposite
a logical indicating whether you want to check not the value with largest difference from the mean, but opposite (lowest, if most suspicious is highest etc.). Default is FALSE

two_sided
Logical value indicating if there is a need to treat this test as two-sided. Default is FALSE

group_by_location
TRUE/FALSE default is FALSE

keep_data_object
TRUE/FALSE default is FALSE