Function to calculate prediction interval

Function to calculate prediction interval

pred_int(df, analysis_result = "analysis_result",
  distribution = "distribution", method = "Bonferroni", n_mean = 1,
  k = 1, pi_type = "upper", conf_level = 0.95)

Arguments

df
df data frame of groundwater data in tidy format

analysis_result
the analysis result column

distribution
the distribution column

method
default is "Bonferroni"

n_mean
n.mean positive integer specifying the sample size associated with the future averages. The default value is n.mean=1 (i.e., individual observations). Note that all future averages must be based on the same sample size.

k
k positive integer specifying the number of future observations or averages the prediction interval should contain with confidence level conf.level. The default value is k=1.

pi_type
character string indicating what kind of prediction interval to compute. The possible values are pi_type="two-sided" (the default), pi_type="upper", and pi_type="lower".

conf_level
a scalar between 0 and 1 indicating the confidence level of the prediction interval. The default value is conf.level=0.95

Examples

data("gw_data")

wells <- c("MW-1", "MW-2", "MW-3", "MW-4")

params <- c("Sulfate, total",
            "Arsenic, dissolved",
            "Boron, dissolved")

background <- lubridate::ymd(c("2007-12-20", "2012-01-01"), tz = "UTC")

# first group data by location, param, and background
# estimate percent less than and distribution
background_data <- gw_data %>%
 filter(location_id %in% wells, param_name %in% params,
         sample_date >= background[1] & sample_date <= background[2]) %>%
  group_by(location_id, param_name, default_unit) %>%
  percent_lt() %>%
  est_dist(., keep_data_object = TRUE) %>%
  arrange(location_id, param_name)

background_data %>%
pred_int(., pi_type = "upper", conf_level = 0.99)

#> [90m# A tibble: 12 x 8[39m
#>    location_id param_name default_unit distribution sample_size   lpl    upl
#>    [3m[90m<chr>[39m[23m       [3m[90m<chr>[39m[23m      [3m[90m<chr>[39m[23m        [3m[90m<chr>[39m[23m              [3m[90m<int>[39m[23m [3m[90m<dbl>[39m[23m  [3m[90m<dbl>[39m[23m
#> [90m 1[39m MW-1        Arsenic, ~ ug/L         Nonparametr~          14  -[31mInf[39m  10   
#> [90m 2[39m MW-1        Boron, di~ mg/L         Nonparametr~          14  -[31mInf[39m   1.37
#> [90m 3[39m MW-1        Sulfate, ~ mg/L         Nonparametr~          14  -[31mInf[39m 117   
#> [90m 4[39m MW-2        Arsenic, ~ ug/L         Nonparametr~          14  -[31mInf[39m  10   
#> [90m 5[39m MW-2        Boron, di~ mg/L         Lognormal             14     0   2.69
#> [90m 6[39m MW-2        Sulfate, ~ mg/L         Nonparametr~          14  -[31mInf[39m 200   
#> [90m 7[39m MW-3        Arsenic, ~ ug/L         Nonparametr~          14  -[31mInf[39m  10   
#> [90m 8[39m MW-3        Boron, di~ mg/L         Lognormal             14     0   2.00
#> [90m 9[39m MW-3        Sulfate, ~ mg/L         Lognormal             14     0 172.  
#> [90m10[39m MW-4        Arsenic, ~ ug/L         Normal                14  -[31mInf[39m  14.3 
#> [90m11[39m MW-4        Boron, di~ mg/L         Normal                14  -[31mInf[39m   1.00
#> [90m12[39m MW-4        Sulfate, ~ mg/L         Nonparametr~          14  -[31mInf[39m 120   
#> [90m# ... with 1 more variable: conf_level [3m[90m<dbl>[90m[23m[39m