This function plots multiple groundwater data time series by location, or constituent.

This function plots multiple groundwater data time series by location, or constituent.

ts_plot(df, x = "sample_date", y = "analysis_result",
  facet_var = "location_id", group_var = "param_name",
  lt_measure = "lt_measure", scale_y_trans = "identity", trend = NULL,
  back_date = NULL, comp_date = NULL, limit1 = NULL, limit2 = NULL,
  short_name = FALSE, pnt = 3, ncol = NULL, ...)

Arguments

df
df groundwater data in tidy format

x
x column for x variable

y
y column for y variable

facet_var
column to facet wrap plots by, default is by location

group_var
column to group plots by, default is by constituent

scale_y_trans
type of transformation to use for y scale. Default is "identity". Built-in transformations include "asn", "atanh", "boxcox", "exp", "identity", "log", "log10", "log1p", "log2", "logit", "probability", "probit", "reciprocal", "reverse" and "sqrt".

trend
trend add trend line to time series plot

back_date
dates for background date range

comp_date
dates for compliance date range

limit1
horizontal line 1

limit2
horizontal line 2

short_name
If TRUE, the constituent name will be abbreviated

pnt
size of points on time series plots

ncol
number of columns

...
parameters passed to get_theilsen